User friendly interface to edgeR for microbiome data

Returns an object of class topTags containing results filtered by FDR < alpha for one comparison. get_edgeR_results_all_levels() returns a list of topTags objects returned by get_edgeR_results() for all levels of the first variable in the formula.

Usage

get_edgeR_results(
  formla,
  pseq = NYC_HANES,
  method = c("BH", "IHW"),
  coef = 2,
  alph = 0.01,
  filtering = method[1] == "BH",
  countMinimum = 8,
  percentMinimumHaveCount = NULL,
  nMinimumHaveCount = 3,
  NA.RM = TRUE
)

get_edger_results_all_levels(formla, ...)

edger_list_to_data.frame(list_models)

get_all_edgeR_models(
  vars,
  varlabels = vars,
  adjusted_for,
  to.data.frame = TRUE,
  ...
)

Arguments

formla

formula. specifies the design matrix used by edgeR::glmFit.

pseq

object of class phyloseq

method

character. Specifies the method of multiple testing correction to apply, either "BH" (Benjamini-Hochberg) or "IHW" (independent hypothesis weighting).

coef

integer. Specifies which linear model coefficient to test (default 2).

alph

numeric. Specifies what FDR level is considered alpha, and only keeps results with FDR less than this number.

filtering

logical. Whether or not to apply pre-filtering.

countMinimum

integer. If filtering==TRUE, the minimum count required for an OTU to be retained. Requires either percentMinimumHaveCount or nMinimumHaveCount.

percentMinimumHaveCount

numeric. If filtering==TRUE, the minimum percentage of samples that must have countMinimum counts for an OTU to be retained.

nMinimumHaveCount

numeric. If filtering==TRUE, the minimum number of samples that must have countMinimum counts for an OTU to be retained.

NA.RM

logical. Whether or not to remove samples with NA values in the sample_data().

...

further arguments passed to get_edgeR_results

vars

character vector, containing names of each independent variable for which you would like a separate model.

varlabels

character vector, containing variable labels for printing

adjusted_for

character vector, containing names of the additional adjustment variables to include in each model

to.data.frame

logical. Whether or not to return a combined data.frame of all model results, rather than the default list.